Added class IDeleteNews, delete method in newsService and delete controller. Also fixed format in json
