+++ src/main/java/com/alkemy/ong/service/abstraction/IDeleteNews.java
@@ -1,0 +1,6 @@
+package com.alkemy.ong.service.abstraction;
+
+public interface IDeleteNews {
+
+  void delete(Long id);
+}
